blackbox exporters source_labels must be unquoted
when setting relabel configs we can not provide an array as to_yaml
converts this to folded. blackbox exporter expets this to be a flow
style array without any quotes.

in puppet we provide the information (e.g.):

    'relabel_configs' => [
      {
        'source_labels' => '[__address__]',
        'target_label'  => '__param_target',
      },
      {
        'source_labels' => '[__param_target]',
        'target_label'  => 'instance',
      },
      {
        'target_label' => '__address',
        'replacement'  => '127.0.0.1:9115',
      },
    ],

within the template we gsub the double quotes which '.to_yaml' adds.