daemon: explicitly pass provider to service
This ensures puppet doesn't incorrectly detect the service provider in some circumstances (e.g. if a systemd unit is disabled).
It also has the side-effect of enabling us to actually meaningfully override the provider (until now, setting $init_style would not propagate to the service, so puppet would try to detect the system provider even if you attempted to pass a different one)

fixes #132