deploy env_file_path on server
Without this change, when the install_method is set to `package` and
init_style is set to `none`, the various `prometheus::server` settings
will never do anything because we stop touching the systemd config
file.

By deploying the env_file_path on the server, we allow the "Debian
package" setup (install_method=package and init_style=none) to
propagate the daemon_flags correctly.

By default, the Debian systemd.service file has this entry:

    EnvironmentFile=/etc/default/prometheus
    ExecStart=/usr/bin/prometheus $ARGS

... which makes it use the ARGS define in the "default" config
file.

This is a subset of #303, which should hopefully pass tests and be
more manageable.

Closes: #323