pass options through env_vars if no control over init files
The logic here is that if init_style is non, then we have no control
over the init files themselves *except* through the environment
file.

Typically, the way this works is there is a ARGS environment variable
that gets sucked into the systemd unit file (or the sysv init file).