check node exporter config file instead of bird in node exporter test
For some reason, the node exporter unit test checks for the
inexistence of the bird_exporter config file. Naturally, it doesn't
find it (why would it?).

I suspect this is an artifact of a copy-paste from the bird exporter,
and what was really meant here was to check for the inexistence of the
node exporter config file.