fix quoting in apache exporter
When using the Debian package to deploy the exporter, like this:

    class { 'prometheus::apache_exporter':
      package_ensure    => $package_ensure,
      service_ensure    => $service_ensure,
      export_scrape_job => true,
      install_method    => 'package',
      init_style        => 'none',
      user              => 'prometheus',
      group             => 'prometheus',
      package_name      => 'prometheus-apache-exporter',
      service_name      => 'prometheus-apache-exporter',
    }

... we end up with the options in
`/etc/default/prometheus-apache-exporter`. It *looks* like it will be
fine: the systemd unit file sucks those in as an environment variable,
and then passes that along to the process. But look closely at what
actually gets generated:

    ARGS="--scrape_uri "http://localhost/server-status/?auto" "

This kind of looks okay if you squint a little, but it actually turns
out it's really, really wrong. What it ends up doing is basically
this:

        apache-exporter --scrape_uri "http://localhost/server-status/?auto "

Notice the trailing slash here. When the exporter makes the request to
Apache next, it ends up doing this query:

    GET /server-status/?auto  HTTP/1.1
    [...]

At first glance, this looks fine: just a regular HTTP request, right?
But if you look more closely, there are *two* spaces before the
HTTP/1.1 Apache doesn't like that, *at all*. It doesn't even wait for
the rest of the headings to come in, it immediately answers with:

    HTTP/1.1 400 Bad Request

... and closes the socket. That, naturally, makes the exporter quite
unhappy and totally unable to collect metrics.

By using single quotes, we work around the quoting issue and correctly
pass the right URL down to the exporter, at least in Debian.